[REF] refactor product_standard_price_tax_included * OCA convention. (folder name, file name, readme, header, ...) [ADD] maintainer and developpement status [FIX] set standard_price_tax_included a company_dependent field
