+++ product_standard_price_tax_included/README.rst
@@ -1,10 +1,29 @@
-.. image:: https://img.shields.io/badge/licence-AGPL--3-blue.svg
-   :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-   :alt: License: AGPL-3
+=================================
+Product - Cost Price Tax Included
+=================================
 
-=======================================================
-Brings a Cost Price Field Tax Included on Product Model
-=======================================================
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fproduct--attribute-lightgray.png?logo=github
+    :target: https://github.com/OCA/product-attribute/tree/8.0/product_standard_price_tax_included
+    :alt: OCA/product-attribute
+.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
+    :target: https://translation.odoo-community.org/projects/product-attribute-8-0/product-attribute-8-0-product_standard_price_tax_included
+    :alt: Translate me on Weblate
+.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
+    :target: https://runbot.odoo-community.org/runbot/135/8.0
+    :alt: Try me on Runbot
+
+|badge1| |badge2| |badge3| |badge4| |badge5| 
 
 In Odoo, the Cost Price Field (standard_price) is by definition 'Tax Excluded'.
 So without this module, the sale price will be bad in this following configuration:
@@ -17,59 +36,93 @@ This module fixes the problem, adding a new field 'Cost Price Tax Included'
 (standard_price_tax_included) on Product Template model, based on Cost Price
 Field and Sale Taxes Setting.
 
-This module create a new 'product.price.type' item, named
+This module create a new ``product.price.type`` item, named
 'Cost Price Tax Included'.
+
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Configuration
+=============
+
+[ This file is optional, it should explain how to configure
+  the module before using it; it is aimed at advanced users. ]
+
+To configure this module, you need to:
+
+#. Go to ...
+
+.. figure:: https://raw.githubusercontent.com/OCA/product-attribute/8.0/product_standard_price_tax_included/path/to/local/image.png
+   :alt: alternative description
+   :width: 600 px
 
 Usage
 =====
 
 User can now set Price List based on this field:
 
-.. image:: /product_standard_price_tax_included/static/description/pricelist.png
+.. image:: https://raw.githubusercontent.com/product_standard_price_tax_included/static/description/pricelist.png
    :width: 100%
 
 The new field is displayed on the product template form:
 
-.. image:: /product_standard_price_tax_included/static/description/product_template.png
+.. image:: https://raw.githubusercontent.com/product_standard_price_tax_included/static/description/product_template.png
    :width: 100%
-
-.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
-   :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/135/8.0
-
-For further information, please visit:
-
-* https://www.odoo.com/forum/help-1
-
 
 Bug Tracker
 ===========
 
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/product-attribute/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
-`here <https://github.com/OCA/product-attribute/issues/new?body=module:%20product_standard_price_tax_included%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/OCA/product-attribute/issues/new?body=module:%20product_standard_price_tax_included%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
+Do not contact contributors directly about support or help with technical issues.
 
 Credits
 =======
 
+Authors
+~~~~~~~
+
+* GRAP
+
 Contributors
-------------
+~~~~~~~~~~~~
 
 * Sylvain LE GAL (https://twitter.com/legalsylvain)
 
-Maintainer
-----------
+Other credits
+~~~~~~~~~~~~~
+
+The development of this module has been financially supported by:
+
+* GRAP, Groupement Régional Alimentaire de Proximité (http://grap.coop)
+
+Maintainers
+~~~~~~~~~~~
+
+This module is maintained by the OCA.
 
 .. image:: https://odoo-community.org/logo.png
    :alt: Odoo Community Association
    :target: https://odoo-community.org
 
-This module is maintained by the OCA.
-
 OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-To contribute to this module, please visit http://odoo-community.org.
+.. |maintainer-legalsylvain| image:: https://github.com/legalsylvain.png?size=40px
+    :target: https://github.com/legalsylvain
+    :alt: legalsylvain
+
+Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
+
+|maintainer-legalsylvain| 
+
+This module is part of the `OCA/product-attribute <https://github.com/OCA/product-attribute/tree/8.0/product_standard_price_tax_included>`_ project on GitHub.
+
+You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
