+++ product_standard_price_tax_included/__init__.py
@@ -1,2 +1,2 @@
-# -*- coding: utf-8 -*-
-from . import model
+# coding: utf-8
+from . import models
