+++ product_standard_price_tax_included/data/product_price_type.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2015 - Today: GRAP (http://www.grap.coop)
+@author Sylvain LE GAL (https://twitter.com/legalsylvain)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<openerp><data>
+
+    <record id="standard_price_tax_included" model="product.price.type">
+        <field name="name">Cost Price Tax Included</field>
+        <field name="field">standard_price_tax_included</field>
+    </record>
+
+</data></openerp>
