+++ product_standard_price_tax_included/demo/account_tax.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2015 - Today: GRAP (http://www.grap.coop)
+@author Sylvain LE GAL (https://twitter.com/legalsylvain)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<openerp><data>
+
+    <record id="account_tax_tax_included" model="account.tax">
+        <field name="name">Sale Tax 15% (Tax Included)</field>
+        <field name="type_tax_use">sale</field>
+        <field name="type">percent</field>
+        <field name="amount">0.15</field>
+        <field name="price_include" eval="True"/>
+    </record>
+
+</data></openerp>
