+++ product_standard_price_tax_included/demo/product_pricelist.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2015 - Today: GRAP (http://www.grap.coop)
+@author Sylvain LE GAL (https://twitter.com/legalsylvain)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<openerp><data>
+
+    <record id="pricelist_price_tax_included" model="product.pricelist">
+        <field name="name">Sale Price List based on Cost Price Tax Included</field>
+        <field name="type">sale</field>
+    </record>
+
+    <record id="pricelist_version_price_tax_included" model="product.pricelist.version">
+        <field name="name">Sale Price List Version based on Cost Price Tax Included</field>
+        <field name="pricelist_id" ref="pricelist_price_tax_included"/>
+    </record>
+
+    <record id="pricelist_item_price_tax_included" model="product.pricelist.item">
+        <field name="name">Sale Price List Item based on Cost Price Tax Included</field>
+        <field name="price_version_id" ref="pricelist_version_price_tax_included"/>
+        <field name="base" ref="standard_price_tax_included"/>
+        <field name="price_discount">0.0</field>
+    </record>
+
+</data></openerp>
