--- product_standard_price_tax_included/demo/product_pricelist.yml
@@ -1,35 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Product - Cost Price Tax Included Module for Odoo
-#    Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-#    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-- !record {model: product.pricelist, id: pricelist_price_tax_included}:
-    name: Sale Price List based on Cost Price Tax Included
-    type: sale
-
-- !record {model: product.pricelist.version, id: pricelist_version_price_tax_included}:
-    name: Sale Price List Version based on Cost Price Tax Included
-    pricelist_id: pricelist_price_tax_included
-
-- !record {model: product.pricelist.item, id: pricelist_item_price_tax_included}:
-    name: Sale Price List Item based on Cost Price Tax Included
-    price_version_id: pricelist_version_price_tax_included
-    base: !eval (ref('standard_price_tax_included'))
-    price_discount: 0.0
