+++ product_standard_price_tax_included/demo/product_template.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2015 - Today: GRAP (http://www.grap.coop)
+@author Sylvain LE GAL (https://twitter.com/legalsylvain)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<openerp><data>
+
+    <record id="product_product" model="product.product">
+        <field name="name">Product With Sale Tax (Tax Included)</field>
+        <field name="uom_id" ref="product.product_uom_unit"/>
+        <field name="uom_po_id" ref="product.product_uom_unit"/>
+        <field name="standard_price">10</field>
+        <field name="list_price">20</field>
+        <field name="taxes_id" eval="[(4, ref('account_tax_tax_included'))]"/>
+        <field name="categ_id" ref="product.product_category_all" />
+    </record>
+
+</data></openerp>
