--- product_standard_price_tax_included/demo/product_template.yml
@@ -1,30 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Product - Cost Price Tax Included Module for Odoo
-#    Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-#    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-- !record {model: product.product, id: product_product}:
-    name: Product With Sale Tax (Tax Included)
-    uom_id: product.product_uom_unit
-    uom_po_id: product.product_uom_unit
-    standard_price: 10
-    list_price: 20
-    taxes_id:
-        - account_tax_tax_included
