+++ product_standard_price_tax_included/demo/res_groups.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (C) 2015 - Today: GRAP (http://www.grap.coop)
+@author Sylvain LE GAL (https://twitter.com/legalsylvain)
+License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+-->
+<openerp><data>
+
+    <record id="base.group_no_one" model="res.groups">
+        <field name="users" eval="[(4, ref('base.user_root'))]" />
+    </record>
+
+    <record id="base.group_sale_manager" model="res.groups">
+        <field name="users" eval="[(4, ref('base.user_root'))]" />
+    </record>
+
+    <record id="product.group_sale_pricelist" model="res.groups">
+        <field name="users" eval="[(4, ref('base.user_root'))]" />
+    </record>
+
+    <record id="account.group_account_manager" model="res.groups">
+        <field name="users" eval="[(4, ref('base.user_root'))]" />
+    </record>
+
+</data></openerp>
