--- product_standard_price_tax_included/demo/res_groups.yml
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Product - Cost Price Tax Included Module for Odoo
-#    Copyright (C) 2015-Today GRAP (http://www.grap.coop)
-#    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-- !record {model: res.groups, id: base.group_no_one}:
-    users:
-        - base.user_root
-
-- !record {model: res.groups, id: base.group_sale_manager}:
-    users:
-        - base.user_root
-
-- !record {model: res.groups, id: product.group_sale_pricelist}:
-    users:
-        - base.user_root
-
-- !record {model: res.groups, id: account.group_account_manager}:
-    users:
-        - base.user_root
