+++ product_standard_price_tax_included/models/__init__.py
@@ -1,2 +1,2 @@
-# -*- coding: utf-8 -*-
+# coding: utf-8
 from . import product_template
