+++ product_standard_price_tax_included/readme/CONFIGURE.rst
@@ -1,0 +1,10 @@
+[ This file is optional, it should explain how to configure
+  the module before using it; it is aimed at advanced users. ]
+
+To configure this module, you need to:
+
+#. Go to ...
+
+.. figure:: path/to/local/image.png
+   :alt: alternative description
+   :width: 600 px
