+++ product_standard_price_tax_included/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Sylvain LE GAL (https://twitter.com/legalsylvain)
