+++ product_standard_price_tax_included/readme/CREDITS.rst
@@ -1,0 +1,3 @@
+The development of this module has been financially supported by:
+
+* GRAP, Groupement Régional Alimentaire de Proximité (http://grap.coop)
