+++ product_standard_price_tax_included/readme/USAGE.rst
@@ -1,0 +1,9 @@
+User can now set Price List based on this field:
+
+.. image:: /product_standard_price_tax_included/static/description/pricelist.png
+   :width: 100%
+
+The new field is displayed on the product template form:
+
+.. image:: /product_standard_price_tax_included/static/description/product_template.png
+   :width: 100%
