+++ product_standard_price_tax_included/tests/__init__.py
@@ -1,2 +1,2 @@
-# -*- coding: utf-8 -*-
-from . import tests
+# coding: utf-8
+from . import test_module
