--- product_standard_price_tax_included/view/view.xml
@@ -1,46 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- ********************************************************************** -->
-<!--Product - Cost Price Tax Included Module for Odoo                       -->
-<!--Copyright (C) 2015-Today GRAP (http://www.grap.coop)                    -->
-<!--@author Sylvain LE GAL (https://twitter.com/legalsylvain)               -->
-
-<!--This program is free software: you can redistribute it and/or modify    -->
-<!--it under the terms of the GNU Affero General Public License as          -->
-<!--published by the Free Software Foundation, either version 3 of the      -->
-<!--License, or (at your option) any later version.                         -->
-
-<!--This program is distributed in the hope that it will be useful,         -->
-<!--but WITHOUT ANY WARRANTY; without even the implied warranty of          -->
-<!--MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the           -->
-<!--GNU Affero General Public License for more details.                     -->
-
-<!--You should have received a copy of the GNU Affero General Public License-->
-<!--along with this program.  If not, see <http://www.gnu.org/licenses/>.   -->
-<!-- ********************************************************************** -->
-<openerp>
-    <data>
-
-<!-- Model: product.template -->
-        <record id="view_product_template_form" model="ir.ui.view">
-            <field name="model">product.template</field>
-            <field name="inherit_id" ref="product.product_template_form_view"/>
-            <field name="arch" type="xml">
-                <field name="standard_price" position="after">
-                    <field name="standard_price_tax_included" groups="base.group_no_one"/>
-                </field>
-            </field>
-        </record>
-
-<!-- Model: product.product -->
-        <record id="view_product_product_form" model="ir.ui.view">
-            <field name="model">product.product</field>
-            <field name="inherit_id" ref="product.product_normal_form_view"/>
-            <field name="arch" type="xml">
-                <field name="standard_price" position="after">
-                    <field name="standard_price_tax_included" groups="base.group_no_one"/>
-                </field>
-            </field>
-        </record>
-
-    </data>
-</openerp>
